requirements.txt
flask
passlib
flask-mysqldb
Flask-WTF
flask_sqlalchemy
flask_cors
flask-jwt-extended
flask-mail